+++ R/ismev.R
@@ -21,6 +21,7 @@
 #' by an error message produced by \code{\link{alogLik}}.
 #' @return An object inheriting from class \code{"chandwich"}.  See
 #'   \code{\link[chandwich]{adjust_loglik}}.
+#'
 #'   \code{class(x)} is a vector of length 5. The first 3 components are
 #'   \code{c("lax", "chandwich", "ismev")}.
 #'   The remaining 2 components depend on the model that was fitted.
@@ -87,6 +88,24 @@
 #'   adj_reg_rain_fit <- alogLik(reg_rain_fit)
 #'   summary(adj_reg_rain_fit)
 #'   }
+#'   # Binomial-GP model -----
+#'
+#'   # Use Newlyn seas surges data from the exdex package
+#'   surges <- exdex::newlyn
+#'   u <- quantile(surges, probs = 0.9)
+#'   newlyn_fit <- gpd.fit(surges, u, show = FALSE)
+#'   # Create 5 clusters each corresponding approximately to 1 year of data
+#'   cluster <- rep(1:5, each = 579)[-1]
+#'   adj_newlyn_fit <- alogLik(newlyn_fit, cluster = cluster, binom = TRUE,
+#'                             cadjust = FALSE)
+#'   summary(adj_newlyn_fit)
+#'   summary(attr(adj_newlyn_fit, "pu_aloglik"))
+#'
+#'   # Add inference about the extremal index theta, using K = 1
+#'   adj_newlyn_theta <- alogLik(newlyn_fit, cluster = cluster, binom = TRUE,
+#'                               k = 1, cadjust = FALSE)
+#'   summary(attr(adj_newlyn_theta, "theta"))
+#'
 #'   # PP model -----
 #'
 #'   # An example from the ismev::pp.fit documentation
@@ -203,7 +222,8 @@ alogLik.pp.fit <- function(x, cluster = NULL, use_vcov = TRUE, ...) {
 
 #' @rdname ismev
 #' @export
-alogLik.gpd.fit <- function(x, cluster = NULL, use_vcov = TRUE, ...) {
+alogLik.gpd.fit <- function(x, cluster = NULL, use_vcov = TRUE, binom = FALSE,
+                            k, inc_cens = TRUE, ...) {
   # List of ismev objects supported
   supported_by_lax <- list(ismev_gpd = "gpd.fit")
   # Does x have a supported class?
@@ -218,12 +238,58 @@ alogLik.gpd.fit <- function(x, cluster = NULL, use_vcov = TRUE, ...) {
   # Set the class
   name_of_class <- names(supported_by_lax)[which(is_supported)]
   class(x) <- name_of_class
+  # If cluster has been supplied then we need to check it's length is correct.
+  # If binom = TRUE then it must have the same length as the raw data x$xdata.
+  # If binom = FALSE then, alternatively, it may have the same length as the
+  # vector of threshold excesses.
+  # We check this and, if necessary, extract the cluster values that correspond
+  # to the threshold excesses.
+  # Save the full cluster vector for use later for the Bernoulli distribution
+  full_cluster <- cluster
+  if (!is.null(cluster)) {
+    clength <- length(cluster)
+    if (binom) {
+      if (clength != x$n) {
+        stop("binom=TRUE: ''cluster'' must be the same length as the raw data")
+      }
+      cluster <- cluster[x$xdata > x$threshold]
+    } else {
+      if (clength == x$n) {
+        cluster <- cluster[x$xdata > x$threshold]
+      } else if (clength == x$nexc) {
+      } else {
+        stop("''cluster'' must have the same length as either the raw or
+             exceedance data")
+      }
+    }
+  }
   # Call adj_object() to adjust the loglikelihood
   res <- adj_object(x, cluster = cluster, use_vcov = use_vcov, ...)
   if (x$trans) {
     class(res) <- c("lax", "chandwich", "ismev", "gpd", "nonstat")
   } else {
     class(res) <- c("lax", "chandwich", "ismev", "gpd", "stat")
+  }
+  # For a stationary model, add the adjusted binomial log-likelihood for the
+  # exceedance probability p_u if binom = TRUE
+  if (!x$trans && binom) {
+    exc <- x$xdata > x$threshold
+    fitb <- fit_bernoulli(exc)
+    afitb <- alogLik(fitb, cluster = full_cluster, ...)
+    attr(res, "pu_aloglik") <- afitb
+    class(res) <- c("lax", "chandwich", "ismev", "bin-gpd", "stat")
+  }
+  # For a stationary model, and if a valid k has been supplied, then calculate
+  # the K-gaps model log-likelihood
+  if (!x$trans && !missing(k) && length(k) == 1) {
+    if (!is.numeric(k) || k < 0 || length(k) != 1) {
+      stop("k must be a non-negative scalar")
+    }
+    # Call exdex::kgaps() to estimate the extremal index theta and store values
+    # from which the log-likelihood can be calculated
+    theta <- exdex::kgaps(data = x$xdata, u = x$threshold, k = k,
+                          inc_cens = inc_cens)
+    attr(res, "theta") <- theta
   }
   return(res)
 }